+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 0.3.0
 
 * Update shiro plugin dependency to 1.2.1
+* Add support for Single Sign Out [#5](https://github.com/commercehub-oss/grails-shiro-cas/issues/5)
 
 # 0.2.0